Libmate/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
Werkzeug==2.3.7
cryptography==41.0.7
schedule==1.2.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0